solving conflicts
